[qmf] Follow the synchronizationEnabled flag when folders are not specified. Contributes to JB#47389
This is to support configurable email folder synchronisation. Applies
the following upstream patch.

Follow the synchronizationEnabled flag when folders are not specified

If the synchronizationEnabled flag is set to false, a variety
of strategy, like search, folder listing, will not search in
such folders. This is maybe a too restrictive usage of this
flag. This patch change the behavior of this flag to act on
synchronization strategies only, when desired.

It activates it only for message retrieval in all folders.