[qmf] Set new IMAP folders to inherit SynchronizationEnabled flag. Contributes to JB#47389
This is to support configurable email folder synchronisation. Applies
the following upstream patch.

When new folders are received from the server, the client-side
SynchronizationEnabled flag must be set for them. Previously they were
always set to true, but if they're inside a folder that isn't being
synced, this probably doesn't reflect the user's intent.

This change makes it so that for IMAP any new folders received from the
server will have the SynchronizationEnabled flag set to whatever value
their parent has it set to. If they have no parent (i.e. are in the root
folder) then the flag is set to true as before.