[qmf] Run qmf-accountscheck with privileges. Contributes to JB#48852
If the accounts database is set to require privileges for read/write
access, then the qmf-accountscheck executable will need to run with
these increased privileges too.

This change sets the service executable to be run by the booster/invoker
and also adds a privilege file so it has the privileges it needs.