[messagingframework] Trigger notification if account credentials are incorrect. Contributes to JB#47900
For accounts that use passwords, if the credentials are incorrect this
change will cause the "CredentialsNeedUpdate = true" property to be set
on the account, triggering signond to show a notification to the user.

This mirrors the behavour of qmf-oauth2-plugin. See
SSOOauth2Plugin::credentialsNeedUpdate() in
base-qmf-oauth2-plugin/oauth2plugin.cpp

This commit applies the change by ammending the previous 0004 patch. It
also removes the numbering from the patches to simplify future patch
creation.