[messagingframework] Add a patch to SSO integration.
This patch simplify SSO patch itself and
will eventually be squashed with the main
SSO patch. It is exposing the username/
password (through the password plugin) to
SMTP, POP and IMAP protocols. It is then
possible to use upstream configuration
objects and code path for login.

This commit also move some parts belonging
to QtAccount integration to the proper patch.
It also takes some code from the KEEPALIVE
patch that are fixup to the SSO integration
patch.