[messagingframework] Remove unnecessary calls to accountsUpdated().
Now that setAccount() in SMTP/POP/IMAP clients is not
used anymore to reset the configuration (since configuration
is reloaded on any new connection), there is no need
to call setAccount() on account change.

One upstream patch forbids this by setting the account
on construction. This simplifies a bit further the existing
patch for Sailfish OS related to libaccount integration.